Add memory handling logic
Memory trait is applied to structs that need to provide read/write logic to and from the bus.
The bus provides a central location where the CPU and other components can talk to each other.
The bus, the CPU, and provided components for the Bus memory map are behind an Arc and RwLock OR Mutex. In order to read or write from the bus, a lock is required to prevent conflict. As well, this allows these components to be thread-safe.